move helper functions to header file
